requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
xgboost==2.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
plotly==5.17.0
sqlalchemy==2.0.23
openpyxl==3.1.2
fpdf2==2.7.6
flask==3.0.0
mixpanel==4.10.0